Extracting functionality into a util class.
